backend/requirements.txt
livekit
livekit-agents
livekit-plugins-groq

langgraph
langchain
langchain-core
langchain-groq
groq

python-dotenv
pydantic

aiohttp
supabase
fastapi
uvicorn
